requirements.txt
fastapi
uvicorn
neo4j
python-dotenv
reportlab
pathlib
openai
